+++ lib/src/backends/litert_lm/litert_lm_backend_web.dart
@@ -1126,6 +1126,9 @@ class LiteRtLmBackend
     if (params.grammarLazy) {
       unsupported.add('grammarLazy');
     }
+    if (params.thinkingBudget != null) {
+      unsupported.add('thinkingBudget');
+    }
     if (params.grammarTriggers.isNotEmpty) {
       unsupported.add('grammarTriggers');
     }