+++ packages/llamadart_llama_cpp_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Updated Apple SwiftPM native pin to `leehack/llamadart-native@b9969-llamadart.1`.
+
 ## 0.0.9
 
 * Updated Apple SwiftPM native pin to `leehack/llamadart-native@b9935`.